Experiment with rewriting tests using RTL.
